Model/simple_/requirements.txt
PyMuPDF==1.26.3
langchain==0.3.15
langchain-core>=0.3.31,<0.4.0
langchain-community==0.3.15
langchain-openai==0.2.5
faiss-cpu==1.8.0
requests>=2.32.3,<3.0.0
numpy>=1.24.0,<2.0
openai>=1.40.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0
pydantic>=2.7.4,<3.0.0
